# Improve to_bibtex performance