@alejandrogallo 2022-12-18
The changes look good! It's not much increase in performance but it can be as you say for bigger libraries one would notice a difference, which I might notice, that's good!