@alexfikl 2022-12-18
> The changes look good! It's not much increase in performance but it can be as you say for bigger libraries one would notice a difference, which I might notice, that's good!

Thanks for taking a look!

I ran `papis --profile` and the remaining time seems to be spent in loading the db and doing the filtering (in this case), so left for another time.